win32: Allow installation as SYSTEM user
This is not a use-case that we recommend, but SCCM users
seem to run into this. The problem with installing it as
SYSTEM is that it can't be updated easily by normal user
accounts.

Calling _SHGetKnownFolderPath fails with an exception for
various user-level variables, so we avoid doing this and
call it lazily and for the appropriate variables only.