1.4.5: Various fixes and workarounds for misconfigured systems.
Exceptions are no longer thrown from knownfolders as they would
prevent the whole top level dir ('system' or 'user') from being
created. Instead, they are passed up and stored and workarounds
are later implemented based on that.

There are two exceptions passed back from knownfolders now. One
for when a folder is not found (the existing one) and a new one
for when a folder could not be verified (seems to equate to it
not existing or not being writable).

Implements an unused general workaround system for problematic
dirs whereby the other one ('system' vs 'user') can be checked
in for ones that do not exist.

A specific workaround is implemented for 'user/documents'. Here
'/user/profile' + 'Documents' is checked. This seems to be safe
even in the face of internationalization.

Finally, a second log stream using OutputDebugStringW is added.
This stream is more verbose (DEBUG level) than the console one.
The idea here is to use Microsoft's DebugView tool when we have
reports of installation problems.