Redo the creation of shortcuts
It was very hard to follow and recent changes meant
that both 'system' and 'script' shortcuts went through
cwp.py. This meant that it was impossible to avoid even
when you wanted to call activate.bat. I have made it so
that 'script' still does this but 'system' does not

Also, I simplified the code and fixed some bugs around
variable expansion and also implemented special handling
of cmd.exe /K or /C which was necesary to allow spaces in
arguments.